fix: Debug모드에서 sol lua 관련 터지는 문제 해결
