+++ EngineSIU/EngineSIU.vcxproj
@@ -183,6 +183,7 @@ xcopy /y /d "$(SolutionDir)$(ProjectName)\Engine\Source\FMOD_API\studio\lib\$(Pl
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="EngineSIU\Engine\Script\LuaBinding.cpp" />
     <ClCompile Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Actors\AnimPlayerActor.cpp" />
     <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderCompileData.cpp" />
     <ClCompile Include="EngineSIU\Engine\Source\Developer\ShaderCompileUtils\ShaderCompile.cpp" />
@@ -324,7 +325,7 @@ xcopy /y /d "$(SolutionDir)$(ProjectName)\Engine\Source\FMOD_API\studio\lib\$(Pl
     <ClCompile Include="GameUI.cpp" />
     <ClCompile Include="SoundManager.cpp" />
     <ClCompile Include="Timer.cpp" />
-    <ClCompile Include="UScriptComponent.cpp" />
+    <ClCompile Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Components\UScriptComponent.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Actors\AnimPlayerActor.h" />
@@ -525,7 +526,7 @@ xcopy /y /d "$(SolutionDir)$(ProjectName)\Engine\Source\FMOD_API\studio\lib\$(Pl
     <ClInclude Include="GameUI.h" />
     <ClInclude Include="SoundManager.h" />
     <ClInclude Include="Timer.h" />
-    <ClInclude Include="UScriptComponent.h" />
+    <ClInclude Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Components\UScriptComponent.h" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="EngineSIU\Font\IconDefs.h" />
