+++ EngineSIU/EngineSIU.vcxproj.filters
@@ -1243,7 +1243,7 @@
       <Filter>EngineSIU\Font</Filter>
     </ClInclude>
     <ClCompile Include="EngineSIU\Engine\Script\LuaManager.cpp" />
-    <ClCompile Include="UScriptComponent.cpp" />
+    <ClCompile Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Components\UScriptComponent.cpp" />
     <ClCompile Include="EngineSIU\Engine\Source\Runtime\Core\Math\BoxSphereBounds.cpp" />
     <ClCompile Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Components\BoxComponent.cpp" />
     <ClCompile Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Components\CapsuleComponent.cpp" />
@@ -1255,6 +1255,11 @@
     <ClCompile Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Engine\Physics\OverlapResult.cpp" />
     <ClCompile Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Engine\Physics\PhysicsSystem.cpp" />
     <ClCompile Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Actors\AnimPlayerActor.cpp" />
+    <ClCompile Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Actors\CameraActor.cpp" />
+    <ClCompile Include="GameUI.cpp" />
+    <ClCompile Include="SoundManager.cpp" />
+    <ClCompile Include="Timer.cpp" />
+    <ClCompile Include="EngineSIU\Engine\Script\LuaBinding.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Natvis Include="EngineSIU\EngineSIU.natvis" />
@@ -1311,7 +1316,7 @@
     </ClInclude>
     <ClInclude Include="EngineSIU\Engine\Script\LuaBinding.h" />
     <ClInclude Include="EngineSIU\Engine\Script\LuaManager.h" />
-    <ClInclude Include="UScriptComponent.h" />
+    <ClInclude Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Components\UScriptComponent.h" />
     <ClInclude Include="EngineSIU\Engine\Source\Runtime\Core\Math\BoxSphereBounds.h" />
     <ClInclude Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Components\BoxComponent.h" />
     <ClInclude Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Components\CapsuleComponent.h" />
@@ -1323,5 +1328,9 @@
     <ClInclude Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Engine\Physics\OverlapResult.h" />
     <ClInclude Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Engine\Physics\PhysicsSystem.h" />
     <ClInclude Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Actors\AnimPlayerActor.h" />
+    <ClInclude Include="EngineSIU\Engine\Source\Runtime\Engine\Classes\Actors\CameraActor.h" />
+    <ClInclude Include="GameUI.h" />
+    <ClInclude Include="SoundManager.h" />
+    <ClInclude Include="Timer.h" />
   </ItemGroup>
 </Project>
