+++ EngineSIU/EngineSIU/Engine/Source/Editor/PropertyEditor/PropertyEditorPanel.cpp
@@ -5,21 +5,24 @@
 
 #include "World/World.h"
 #include "Actors/Player.h"
+
 #include "Components/Light/LightComponent.h"
 #include "Components/Light/PointLightComponent.h"
 #include "Components/Light/SpotLightComponent.h"
 #include "Components/Light/DirectionalLightComponent.h"
+#include "Components/ProjectileMovementComponent.h"
 #include "Components/Light/AmbientLightComponent.h"
 #include "Components/StaticMeshComponent.h"
+#include "Components/HeightFogComponent.h"
 #include "Components/TextComponent.h"
+#include "Components/UScriptComponent.h"
+
 #include "Engine/EditorEngine.h"
 #include "Engine/FLoaderOBJ.h"
 #include "UnrealEd/ImGuiWidget.h"
 #include "UObject/Casts.h"
 #include "UObject/ObjectFactory.h"
 #include "Engine/Engine.h"
-#include "Components/HeightFogComponent.h"
-#include "Components/ProjectileMovementComponent.h"
 #include "GameFramework/Actor.h"
 #include "Engine/AssetManager.h"
 #include "UObject/UObjectIterator.h"
