+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Level.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "UObject/Object.h"
 #include "UObject/ObjectMacros.h"
-#include "../../../../../../UScriptComponent.h"
+
 
 class AActor;
 class UWorld;
