+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Renderer/Renderer.h
@@ -63,6 +63,7 @@ public:
     //==========================================================================
     // 렌더 패스 관련 함수
     //==========================================================================
+    void TickRender(float DeltaTime);
     void Render(const std::shared_ptr<FEditorViewportClient>& Viewport);
     void RenderViewport(const std::shared_ptr<FEditorViewportClient>& Viewport); // TODO: 추후 RenderSlate로 변경해야함
