+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Windows/D3D11RHI/GraphicDevice.cpp
@@ -284,6 +284,24 @@ void FGraphicsDevice::CreateAlphaBlendState()
     {
         MessageBox(NULL, L"AlphaBlendState 생성에 실패했습니다!", L"Error", MB_ICONERROR | MB_OK);
     }
+
+    blendDesc = {};
+    blendDesc.AlphaToCoverageEnable = FALSE;
+    blendDesc.IndependentBlendEnable = FALSE;
+    blendDesc.RenderTarget[0].BlendEnable = FALSE;
+    blendDesc.RenderTarget[0].SrcBlend = D3D11_BLEND_SRC_ALPHA;
+    blendDesc.RenderTarget[0].DestBlend = D3D11_BLEND_INV_SRC_ALPHA;
+    blendDesc.RenderTarget[0].BlendOp = D3D11_BLEND_OP_ADD;
+    blendDesc.RenderTarget[0].SrcBlendAlpha = D3D11_BLEND_ONE;
+    blendDesc.RenderTarget[0].DestBlendAlpha = D3D11_BLEND_ZERO;
+    blendDesc.RenderTarget[0].BlendOpAlpha = D3D11_BLEND_OP_ADD;
+    blendDesc.RenderTarget[0].RenderTargetWriteMask = D3D11_COLOR_WRITE_ENABLE_ALL;
+
+    hr = Device->CreateBlendState(&blendDesc, &OpaqueBlendState);
+    if (FAILED(hr))
+    {
+        MessageBox(NULL, L"OpaqueBlendState 생성에 실패했습니다!", L"Error", MB_ICONERROR | MB_OK);
+    }
 }
 
 void FGraphicsDevice::ChangeRasterizer(EViewModeIndex ViewModeIndex)
