+++ EngineSIU/EngineSIU/Shaders/SlateShader.hlsl
@@ -44,7 +44,6 @@ PS_Input mainVS(uint VertexID : SV_VertexID)
 
 float4 mainPS(PS_Input Input) : SV_TARGET
 {
-    float4 Scene = ViewportTexture.Sample(Sampler, Input.UV);
-
-    return Scene;
+    float4 SceneColor = ViewportTexture.Sample(Sampler, Input.UV);
+    return SceneColor;
 }
